gplazma2-kpwd: Fix leak of confidential information
The kpwd gplazma2 plugin wraps a UserPwdRecord into a Principal. This 
results in the hashed password being leaked into the public part of a 
Subject, thus resulting in the hashed password appearing in log files, 
WebDAV door HTML renderings, etc.

This patch fixes the plugin such that the UserPwdRecord is no longer 
wrapped in a Principal. Instead the required fields are copied to 
KpwdPrincipal.

UserPwdRecord is also updated such that in no longer includes the hash 
in the toString() output. This is a defensive move to protect against 
the hash showing up in log files.

The patch cannot be back ported as is as it affects the serialized 
format of KpwdPrincipal. A separate patch for 2.0 and 1.9.12 has been 
submitted (http://rb.dcache.org/r/4004/).

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Acked-by: Jan Schaefer <jan.schaefer@desy.de>
Patch: http://rb.dcache.org/r/3992/