Webadmin: Increase Session timeout
During testing on Dimas(Desy) instance his session timed out quickly - which is strange, because on my machine the last patch about a year ago solved this, by setting the timeout in the webdefault.xml of jetty to a higher value. Now I redo this in the web.xml of the wicket application.

patch: http://rb.dcache.org/r/4019/

acked-by: Gerd

target: trunk
request: 2.0