libs: Upgrade to latest version of EMI voms and trustmanager
I uploaded the latest versions of trustmanager, trustmanager-axis and
vomsjapi to the dCache Nexus repository.

It seems the glite-security-util-java package is no longer needed (it
appears to duplicate some of the classes also found in vomsjapi,
although in a different Java package; my guess is that it is an old
version of vomsjapi and trustmanager).

Target: trunk
Ticket: http://rt.dcache.org/Ticket/Display.html?id=6858
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/4037/