info-provider: add support for publishing the EMI version
EMI requires us to publish, within the GLUE 2.0 DIT, some EMI 
information.  This information is that the distribution is EMI and 
which version of EMI is installed.

The full requirement is to gather this information and publish a 
dCache version and EMI version for each door and a single dCache 
version and EMI version for the overall dCache instance.  The EMI 
version should come from the contents of the /etc/emi-version file.

This patch adds support for publishing a single EMI version based on 
the /etc/emi-version file on whichever machine runs the info-provider.  
This information is published as two additional attributes:

    GLUE2EntityOtherInfo=DistributionName=EMI
    GLUE2EntityOtherInfo=DistributionVersion=<contents of /etc/emi-version>

If the file /etc/emi-version is not found (or isn't readable) then 
these two attributes are not published.

This patch also syncs dCache's copy of Xylophone with the latest 
version from github.  The additional functionality was introduced to 
support publishing these attributes and unit-tests were written to 
verify they work correctly.

Target: trunk
Request: 2.1
Patch: http://rb.dcache.org/r/4061/
Acked-by: Gerd Behrmann