scripts: fix dcache script for Solaris machines
Patch http://rb.dcache.org/r/3976/ (committed as trunk@16683) 
enhanced the "dcache status" command by providing a return-code that 
matches dCache's state.

Unfortunately, this patch used some features that are under-specified 
in the POSIX standard.  This is sufficient for the script to fail 
when run under the Solaris POSIX shell.

The issue is with how the shell handles an arithmetic expansion that 
references a variable with using the usual ${} expansion.  POSIX 
notes that $((foo)) should give the same result as $((${foo})) but 
doesn't specify how $((foo+1)) should be handled.  With BASH, this is 
always treated as equivalent to $((${foo}+1)).  With Solaris POSIX 
shell the behaviour depends on whether a variable foo is defined: if 
defined then $((foo+1)) is treated as $((${foo}+1)), if not then it's 
treated as $((${foo+1})).

This patch fixes the shell in two ways: by making the references 
explicit and by making sure variables are defined.  Making references 
explicit, $((${foo}+1)), is preferred since POSIX-compliant shells 
will expand such statement correctly.  For the rc_0, rc_1 and rc_3 
counters, the patch makes sure they are defined.  This is because 
explicit double-deferencing variables within an arithmetic expansion 
doesn't seem to work.

Target: trunk
Request: 2.1
Request: 2.0
Request: 1.9.12
Patch: http://rb.dcache.org/r/4073/
Acked-by: Gerd Behrmann