pool: Add pool convert subcommand to dcache script
The procedure to convert between meta data stores is rather obscure and
involves setting an import repository, restarting the pool and then
clearing the import repository. This has confused lots of users.

This patch adds a 'dcache pool convert' command to convert from the
current meta data store to another format. The patch also implements
conversion from the Berkeley DB to the file based meta data store. This
is in particular useful when downgrading a pool to a version with an
older, incompatible, Berkeley DB format. It is now possible to first
convert the meta data to the file based format, downgrade, and then
convert the meta data to the Berkeley DB based format.

The old mechanism is removed.

RELEASE NOTES: metaDataRepositoryImport is not longer supported and 
marked as a forbidden property. The 'dcache pool convert' command 
replaces the functionality.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/4075/