billing: fix bugs introduced with recent patch
Unfortunately patch http://rb.dcache.org/r/3955/ (committed as 
trunk@16794) introduced a number of bugs:

    o billing service requires billing database to exist, even if 
      logging billing to db is disabled,

    o httpd service requires billing database to exist, even if 
      logging billing to db is disabled,

    o billing doesn't work as the messageArrived method has wrong 
      signature for UniversalSpringCell,

    o confusion between billingDatabaseHost and billingDbHost 
      properties in defaults file,

    o change of default value for billingDisableTxt (formally: false, 
      current: true),

    o confused sense of billingDisableTxt in code (setting 
      billingDisableTxt=true enables output),

    o non-InfoMessage messages sent to billing would result in NPE,

    o if text logging is enabled, the error log files are sent all 
      messages not just the errors,

    o if text logging is disabled, all messages are logged to a single 
      file: 'billing-error-null'.

Mostly the fixes for these items are fairly straight-forward.

This patch also updated some of the comments in the defaults file, 
which had become inaccurate.

One point to note is that, in fixing the messageArrived method, 
billing is limited to accepting only InfoMessage (and subclasses 
thereof).  The original code accepted an arbitrary CellMessage 
envelope and would treat those with an InfoMessage content specially.  
From my reading of dCache code, billing is only sent messages that are 
subclasses of InfoMessage (in fact, all messages are a subclass of 
PnfsFileInfoMessage), so the code to handle non-InfoMessage case was 
redundant.

Target: trunk
Request: 2.1
Patch: http://rb.dcache.org/r/4070
Acked-by: Tigran Mkrtchyan
Acked-by: Jan Schaefer
Acked-by: Gerd Behrmann
Acked-by: Albert Rossi