configuration: add a debug layout printer
The dCache configuration support includes the ability to print a 
dCache configuration in different formats.  Currently available are a 
couple of different formats: as the definition for a POSIX shell 
function (the "shell oracle") and as a set of XML entity definitions.

Both formats must encode the keys and values so that reading the 
output can be difficult.  This makes it hard to investigate problems 
with the configuration system.

This patch adds support for a 'debug' format.  The following command 
illustrates how to generate a dCache configuration in this format:

    [/opt/d-cache/bin/]dcache loader compile -debug

The debug format is intended for debugging configuration problems by 
providing all configuration information in a format that humans should 
find relatively easy to read.  The format is similar to (but distinct 
from) the dCache layout file format.  One noticeable addition is that, 
if a property value contains references to other property values then 
both the unexpanded and expanded values are printed (the expanded 
version is printed as a comment).  The output also includes the domain 
and service declarations.

The output is distinct from dCache configuration since property values 
are not escaped.  The aim is to improve readability at the expense of 
the output being lossy if the value contains any special characters.

Target: trunk
Patch: http://rb.dcache.org/r/4080/
Acked-by: Tigran Mkrtchyan