maven: add missing Spring tx dependencies
A couple of projects (dcache-core and dcache-chimera) make explicit use 
of Spring's transactional support but this isn't mentioned in these 
projects' pom.xml files.  For dcache-core, its dependency on spring-tx 
is satisfied transitively either via the project's dependency on 
spring-orm or on spring-aspects; for dcache-chimera, the dependency is 
satisfied via it's dependency on dcache-core.

This patch makes these projects' dependency on spring-tx explicit.  
Although I believe this isn't strictly needed (due to transitive nature 
of dependencies), since the code makes explicit reference to spring-tx 
classes, it seems reasonable.

As a side-effect (and main motivation) of this patch, netbeans is able 
to understand the code and its dependencies better.

Target: trunk
Patch: http://rb.dcache.org/r/4085/
Acked-by: Gerd Behrmann