httpd: enable /plots alias only when plotting is enabled
The new billing plot engine stores plot information in a directory 
(/var/dcache/plots by default), which we don't ship.  This means 
that, when starting up httpd, the log file will always emit a 
warning that this directory doesn't exist.

This patch provides this alias only when billing is enabled.

In the future, there should probably be an additional patch that 
ensures dCache is shipped with this as an empty directory.

Target: trunk
Patch: http://rb.dcache.org/r/4089/
Acked-by: Tigran Mkrtchyan