configuration: use 0 to indicate random port for cells topology discovery
Current trunk (and 2.1) uses the keyword 'random' to indicate that a 
randomly chosen UDP port should be used.

Gerd reminded me that it is common to use '0' to represent this 
situation; amongst other uses, this is in keeping with how port 
numbers are configured for pools undertaking a passive dcap transfer.

This patch updates dCache so it accepts '0' instead of 'random'.  It 
also adds explanatory text, so that the information is obvious to 
those who don't know the convention.

Target: trunk
Request: 2.1
Patch: http://rb.dcache.org/r/4093/
Acked-by: Gerd Behrmann