bootloader: Move CLI logback config to submodule
For several commands the logback configuration was embedded in the code.
For other commands a special logback-cli.xml was referenced on the
command line. For the remaining commands a non-existing log4j file was
referenced.

This patch adds the logback-config submodule. It contains a logback
configuration suitable for use for CLI utilities. The configuration is
parameterized such that the log level can be adjusted.

This configuration is now used for all the CLI utilities. It is also
used by the dCache boot loader until the real logback configuration is
loaded.

Since the patch fixes the references to non-existing log4j setup, I
submit the patch for inclusion in 2.1.

Target: trunk
Request: 2.1
Acked-by: Jan Shaefer <jan.schaefer@desy.de>
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/4103/