configuration: fix annotations for properties with '+' in their name
Gerd discovered that some property references were not expanded 
correctly.

After some investigation, the problem was triggered by properties that 
both included a '+' in the property name and were annotated.  The cause
was that, with a '+' in the property name, the annotation was not
processed correctly and the annotation declaration remained in the
property name.  This resulted in property references failing to expand.

This patch adjusts the property handling so that properties may have 
any valid characters.

The problem is observable in Trunk and 2.1; however, I've also put in 
request for 2.0 and 1.9.12 since these branches will experience similar
problems if, at some future time, these branches include any properties
that are both annotated and has a name with a character not in [-\w./].

Target: trunk
Request: 2.1
Request: 2.0   
Request: 1.9.12
Acked-by: Gerd Behrmann 
Patch: http://rb.dcache.org/r/4099/