billing: fix XML to allow to-DB to startup
Billing current doesn't start up if logging to a database is enabled.  
This is because the Spring XML file, in this case, is broken.

This patch fixes this problem.  It simply removes the load-time-weaver 
AspectJ annotation since it isn't needed: unlike pin-manager, billing 
doesn't use AspectJ to weave transaction-management code.

Note that this patch is basically the same as patch 
http://rb.dcache.org/r/4098/

Target: trunk
Request: 2.1
Patch: http://rb.dcache.org/r/4130/
Acked-by: Jan Schaefer