billing: disable plotting if storage directory is missing
Billing's plotting machinery requires a directory in which it will 
store plots of time series.  Current billing does not handle the 
absense of this directory, logging many stacktraces for each attempt at 
plotting.

This bad behaviour in the plotting code needs to be fixed (a later 
patch), but this patch simply checks whether the directory used to 
store plot files exists.  If it doesn't exist then an error message is 
printed that describes how an administrator can correct the situation.  
It also disables the plotting machinery since, without the storage 
directory, this cannot work.

Target: trunk
Request: 2.1
Patch: http://rb.dcache.org/r/4136/
Acked-by: Gerd Behrmann