billing: fix another race-condition
There is another problem with billing's transaction-committer code, at 
least, as it is being tested.  The "bug" will (very likely) not cause 
a problem in production, but it interferes with our ability to 
accurately test our code.

When the TimedCommitter thread starts, it first tries to commit a 
transaction before sleeping for the configured duration.

This behaviour introduces a race between the TimedCommitter thread's 
initial transaction-commit and the initial burst of writes done by the 
testDelayedCommit unit test.

On machines to which I have access, testDelayedCommit wins this race 
and manages to get some (~12) writes in before the TimedCommitter's 
initial transaction-commit.

This patch fixes the problem simply by ensuring there is a sleep 
before the first transaction-commit.

Target: trunk
Request: 2.1
Patch: http://rb.dcache.org/r/4135/
Acked-by: Tigran Mkrtchyan