Potential memory leak from missing close() in DataNucleusBillingInfo.
A small regression in the implementation of the DAO get() method: the finally{} clause should have included a close() of the PersistenceManager (as do the remove() methods), but did not.

If the persistence manager's close() is not called, the manager is not released and thereby removed from the factory's collection, and in turn all the internal memory of the manager is also not garbage-collected.  This causes an OutOfMemoryError after enough reads have accumulated.

The close() call has been added, but we have also done the following to ensure better scaling:

1.  The implementation of the delayed commit of inserts used to make a fresh transaction object for each batched insert; it now closes and recreates the PersistenceManager as well, thus paring down the footprint of the factory even further.  Synchronization is done on the factory object instead of the manager (since it can now be null).
2.  Three lines have been added to the utility method for creating a Query object; these are suggested optimizations which avoid some lower level caching and which should also reduce overall footprint.

Patch: http://rb.dcache.org/r/4145/
Acked-by: Dmitry
Target: trunk
Request: 2.1