Plots filter incorrectly includes error code != 0
The hourly plots generated for bytes and transfers from the fine-grained tables should not include failed reads and writes.  The triggers which produce the coarse-grained tables correctly exclude operations with error code != 0, but the query filters for the hourly plots neglected to specify this.

"&& errorCode == 0" has been added to the two filters in question.

Patch: http://rb.dcache.org/r/4146/
Acked-by: Dmitry
Target: trunk
Request: 2.1