billing: shutdown billing cell cleanly
The billing cell accesses the database to write accounting information 
(if so configured).  This access includes the delayed-commit thread, 
which periodically commits the outstanding transaction.

There is currently invocation of the support for shutting down this 
thread when dCache shuts down.  This leads to dCache complaining that 
threads are still active.

This patch adds the missing connection so that BillingCell knows when 
to shutdown and uses this knowledge to kill the delayed-commit thread.

Target: trunk
Request: 2.1
Patch: http://rb.dcache.org/r/4141/
Acked-by: Tigran Mkrtchyan