http: fix error logging
If the httpd service contains aliases that map URI elements to a 
directory or file then requests for that URI will return the contents 
of that directory (rendered as HTML) or file.  If there is a problem 
fetching the directory or file then a stack trace is logged.

A problem fetching a resource is likely to to misconfiguration; logging 
a stack-trace is overkill.

This patch alters the logged message to be a single line.  It includes 
the requested URI and the error message.

Target: trunk
Patch: http://rb.dcache.org/r/4140/
Acked-by: Tigran Mkrtchyan