Webadmin: In case of Exception the PageCache is resetted now
When a component goes down, webadmin shows the last known values for some pages, this behavior is changed, because it makes it easier to recognise, if a component went down. Now in case of exception (most likely these are NoRouteToCell or Timeout Exceptions), the cache is emptied.

patch: http://rb.dcache.org/r/4151/

acked-by: Tigran

target: trunk