Webadmin: Change War-file unpacking behavior.
Antje had a running webadmin that was installed before christmas and left on the vm switched on. Today it did not work anymore. What happend was the following: Jetty unpacks the war-file into the /tmp/.. directory. Since it was not in use for a longer period of time the OS deleted the directory. After that webadmin did not work anymor. By setting the extractWar value to false jetty still unpacks the war-file into tmp, but if it doesn't find anything there anymore it looks the missing content up directly in the war-file. Of, course this is a little less performant, but a nice fallback to prevent complete breakdown of the webapp itself. In future a switch away from the war-file might be considerable and putting the whole webapp into a fixed folder (would also allow the users to customize css files, if wanted). For now this small change should be sufficient.

patch: http://rb.dcache.org/r/4129/

acked-by: Tigran

target: trunk
request: 2.1
request: 2.0
request: 1.9.12