info: add domain-specific information discovered during start-up
This patch adds support for domain-specific information that is 
discovered during a domain's startup.

The information is stored in a specific domain context 'info.static' in 
the form of a sequence of zero or more lines.  Each line describes a 
metric that is added to this domain's part of the info tree (i.e., the 
path "domains.<domain name>").  The metrics are all added under the 
branch 'static'.

The lines that describe a metric have a simple format.  Each non-empty 
line represents some metric to add.  The line describes the metric 
type, the location of the metric and the metric's value.  The metric's 
location uses the dot character as a separator between successive path 
elements; for example:

    S:path.to.metric:metric value

In principal, a metric of any supported type may be added.  This patch 
is limited to supporting Strings metrics; however, adding additional 
types is trivial.

The patch also adds support for creating the info.static context for 
each domain on the startup of this domain.

To illustrate the process the patch includes support for discovering 
whether the /etc/emi-version file exists.  If it does, the contents of 
this file is stored as the metric domains.<domain 
name>.static.environment.distribution.emi-version.  Note that the 
current implementation only if the file contains, at most, a single 
line.  This is the stated format of this file.

Target: trunk
Request: 2.1
EMI-ticket: https://savannah.cern.ch/task/?22269
Patch: http://rb.dcache.org/r/4181/
Acked-by: Tigran Mkrtchyan