info-provider: publish EMI version for doors; use only EMI version from info
Update publishing of doors in GLUE 2.0 to include the EMI version 
(contents of /etc/emi-version).  This information is taken from the 
static information for the domain hosting the door.  If the file 
/etc/emi-version is present on the node hosting the door then two 
attributes are published for the door:

    GLUE2EntityOtherInfo: MiddlewareName=EMI
    GLUE2EntityOtherInfo: MiddlewareVersion=<value from /etc/emi-version>

If /etc/emi-version is not present on the node hosting the door then 
these two attributes are not published.

The patch also adjusts how corresponding information is published on 
the summary object (GLUE2StorageService).  Currently, the info-provider 
injects the contents of /etc/emi-version (on whichever node it is run) 
into the set of dCache configuration properties presented to the 
info-provider.

This is undesirable since the host running info-provider may not run 
any dCache services, so the value of /etc/emi-version is 
unrepresentative.  It also publishes only one EMI version, when a 
dCache instance may be running using several EMI versions.

This patch updates the value published so that the "MiddlewareVersion" 
attributed is repeated for each distinct EMI version that is installed.  
If none of the nodes in a dCache instance has /etc/emi-version then 
neither the "MiddlewareName" nor the "MiddlewareVersion" attribute is 
published.

Since the old method of fetching the contents of /etc/emi-version on 
the info-provider host is no longer used, this patch removes support 
for providing that information as a dCache property.

Note that this patch depends on http://rb.dcache.org/r/4181/

Target: trunk
Request: 2.1
EMI-ticket: https://savannah.cern.ch/task/?22269
Patch: http://rb.dcache.org/r/4183/
Acked-by: Tigran Mkrtchyan