tests: Introduce support for Mockito and Hamcrest matchers
Mockito is one of many mocking libraries on the market. These
libraries make it easy to generate scripted stubs for interfaces
and classes for the purpose of testing. We have several attempts
and semi generic mocking classes in dCache and I think it makes
sense to replace these with one of the standard tools on the
market.

I had a look at the various libraries and Mockito seemed to have
the biggest user community and the nicest syntax. JMockit is an
alternative. While JMockit is clearly much more powerful (allows
mocking of final methods, static members, singletons, etc), the
syntax takes more time to get used to. The limitations of Mockito
also tend to force you to write better code (although that is
debatable - the author of JMockit thinks the other tools force
him to write worse code).

In the end I think we need more readable tests more than the byte
code magic of JMockit, so I went for Mockito. The frameworks can
coexist, so the day we need more powerful tools we can still
import JMockit or PowerMock.


As a demo of Mockito I updated ConsistentStoreTest. Not only is
it now using Mockito, the tests have been rewritten to follow the
style of behavioural driven development. That is, test names are
'shouldDoThisAndThat()' (refering to the class under test
[ConsistentStore in this case]), and tests are structured
according to the 'given, when, then' pattern, ie. 'given these
conditions, when we do this, then these other conditions are true
afterwards'. This tends to give very readable well structured
tests.


I also switched all assertions to the assertThat() method from
JUnit 4.4. These work with Hamcrest matchers and allow very nice
assertions to be written. Even more important is that the error
messages generated when assertions fail are very readable. JUnit
and Mockito come with the core Hamcrest matchers, but I added an
import for the full Hamcrest library (it allows assertions on
collections to be expressed, eg assertThat(collection,
hasItem("bla")) [and yes, assertTrue(collection.contains("bla"))
is the same test, but the error message from the latter is simply
"Expected true, but was false" whereas the former is something
along the lines of "collection was expected to contain item "bla"
but it did not"].


The ConsistentStoreTest class didn't become shorter in terms of
number of lines. This is partly because I added lots of comments
and partly because I am testing more post-conditions than the old
tests. I thought for a long time whether to leave the comments or
not - on one hand tests should be easy to read, making the
comments unnecessary; on the other hand I do think that a
straightforward english description of the requirement adds
value, so I left them in. Notice how they are always written in
the "Given this and given that, when we do this, then condition A
and B are true". The comments don't explain the code - they state
the requirement in plain english.


Due to the additonal post-conditions I discovered that the
Chimera HFile class doesn't implement the delete method. The
patch adds such an implementation.


Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/4184/