Remove gPlazma1 dependencies from gPlazma2.
Several of the current gPlazma2 plugins (X509, GridMap, VoroleMap, VOMS) use utility classes and methods which are contained in the gPlazma1 module.

In anticipation of the eventual deprecation/elimination of gPlazma1, this patch breaks those dependencies by:

1.  Moving selected methods from gplazma.authz.util.X509CertUtil to org.dcache.gplazma.util.CertificateUtils;
2.  Copying gplazma.authz.util.NameRolePair to org.dcache.gplazma.util.NameRolePair;
3.  Copying gplazma.authz.plugins.gridmapfile.GridMapFile to org.dcache.gplazma.util.GridMapFile.

The call-site imports in the gPlazma2 plugins, as well as in an srm and javatunnel class, have been changed to the gPlazma2 equivalents.

NOTE:  there are still further dependencies on gPlazma1 code scattered throughout the code base.  These will be addressed in a later patch.  The dependencies that are changed here are primarily in order to separate gPlazma2 cleanly from gPlazma1.


Patch: http://rb.dcache.org/r/4179/
Acked-by: Tigran
Target: trunk