http: make engines lifecycle aware
The http engines largely do all their work will the thread 
processing a request; however, some engines also include an element 
of background work.  Examples of such engines are 
HttpPoolMgrEngineV3, BillingHistory and HttpBillingHistoryEngine.  
Such engines typically start a thread in their constructor that 
undertakes the background work.

Apart from problems with subclassing, there's no mechianism for the 
engines to learn that dCache is being shutdown.  This is 
undesirable as, without this, the fall-back method of explicitly 
killing threads is used if a cell doesn't shutdown completely 
within a certain time limit.  Apart from the unnecessary delay in 
dCache shutting down, if too many cells need to be killed then the 
supervisor daemon will kill the domain out-right, resulting in an 
unclean shutdown.

dCache provides some hint when a cell is not shutting down cleanly 
by logging both an error message and listing the threads that are 
still running.

This patch adds two events to an HTTP Engine: startup and shutdown.  
The startup method is called after construction and environment 
injection: once the engine should be in a state to run.  The 
shutdown method is called when the engine is expected to stop any 
and all threads it has started.

startup and shutdown methods are added to all the engines.  Most of 
them are no-op but, for those engine that start a thread, the 
thread start is moved to the startup method.  The shutdown method 
kills off the thead and any other engine-initiated activity.

This patch also refactors the AliasEntry 'type' strings so that 
constants are used rather than string-literals.

Target: trunk
Patch: http://rb.dcache.org/r/4142
Acked-by: Tigran Mkrtchyan