info: fix "cannot find StateChangeSet" error message
The info service processes changes by first converting the list of 
changes into a tree-like set of diffs and then walking this tree 
applying the diffs.  One optional part of this list of changes is to 
purge the subtree below a certain point.

A recent patch (http://rb.dcache.org/r/4181/, committed as r.16990) 
uncovered a bug; however, it's possible that this bug is also triggered 
under normal circumstances.  When the bug is triggered, info detected 
an inconsistency and reports this as a message like:

    cannot find StateChangeSet for path domains.nfsserver. Something must have gone wrong.

What is happening is the purge request is for a non-existent part of 
the info tree.  The diffs (StateChangeSet objects) contain one mentions 
a child object that should be iterated into yet, because this object 
doesn't contain the next path-element, the corresponding StateChangeSet 
object doesn't exist.  For example, if the change is to purge the 
subtree starting at a.b.c but the current state contains only 'a' and 
'b' (no 'c') then the StateChangeSet of 'b' doesn't exist.

The same problem can exist if a branch contains no children: if a 
change purges path a.b.c and 'a', 'b' and 'c' all exist but 'c' has no 
children.

This patch fixes the problem.  It also adds a unit test to catch this 
problem and updates an existing test that was actually broken.

Target: trunk
Request: 2.1
Request: 2.0
Request: 1.9.12
Request: 1.9.5
Patch: http://rb.dcache.org/r/4205/
Acked-by: Karsten Schwank