Add gPlazma2 XACML plugin module.
gPlazma1 provided SAML and XACML plugins, but these are missing from gPlazma2.  

This patch brings XACML functionality into v2.  This is essentially the same as what was in gPlazma1. 

An example of a viable gplazma.conf stack:

-----
auth required xacml "gplazma.xacml.service.url=https://gums.fnal.gov:8443/gums/services/GUMSXACMLAuthorizationServicePort"
map required authzdb
session requisite authzdb
-----

The same dependencies on the privilege-xacml and opensaml libraries continue to pertain.

As with the gPlazma1 implementation, there is local in-memory caching of mappings already retrieved from a XACML service; the time-to-live for these tokens is configurable.

Also as with gPlazma1, VOMS validation is turned off by default (but could be activated if so desired).

For extensibility, and primarily for stand-alone testing, the privilege XACML client interface has been exported (IMapCredentialsClient); this is the API which the plugin uses internally; the default (privilege MapCredentialsClient) implementation is provided by the PrivilegeDelegate; the interface is also implemented by the XACMLTestClient, which emulates the mapping service by reading a definition file (test.fixture); this client is used by the Junit tests in XACMLPluginTest.

Notes:  

1) this patch adds a new module to the build (gplazma2-xacml)

2) this patch depends on the previous application of:
       - http://rb.dcache.org/r/4179 (eliminates gPlazma1 dependencies) [committed as revision 17004]

3) there are no unit tests in this patch as we are still trying to decide how best to test functionality that depends on the presence of CA certs and proxies.  A later patch will add the tests, if appropriate.

Patch: http://rb.dcache.org/r/4180
Target: trunk
Ack-ed by: Tigran, Paul