Fixed NSSwitch root group membership test for Mac OS X.
The NSSwitch root group membership test failed on Mac
OS X as there is no group called 'root'. Instead the
default group of root is called 'wheel'. So this version
of the test checks if root is a member of at least one
of the groups.

Patch: http://rb.dcache.org/r/4226
Acked-by: Tigran Mkrtchyan
Target: trunk