spring: add support for bean profiles
Spring v3.1 introduced the concept of bean profiles.  A bean profile 
is some arbitrary string label.  A spring environment (for us, 
something associated with an XML file) has zero or more bean profiles 
active.

When defining beans in an XML file, one or more beans may be placed in 
a container (a 'beans' element) that controls whether they are 
instantiated.  The container has one or more bean profiles as a 
comma-separated list (the 'profile' attribute).  If any of the listed 
bean profiles are active then the beans within the container are 
instantiated; so

    <beans profile='foo,bar'>
        <!-- list of bean defns here -->
    </beans>

may be read as: list of beans to instantiate if profile 'foo' OR 
profile 'bar' are active.

Here's an example:

    <beans>
        <bean id="a"/>

        <bean id="b"/>

        <beans profile="foo">
            <bean id="c"/>

            <bean id="d"/>
        </beans>

        <beans profile="bar">
            <bean id="e"/>
        </beans>

        <beans profile="foo,bar">
            <bean id="f"/>
        </beans>
    </beans>

Without any bean profiles active, only two beans are instantiated (a & 
b).  If profile 'foo' is active then there are five beans (a, b, c, d 
& f).  If bean profile 'bar' is active the four beans are created (a, 
b, e & f).  All six beans are active if profiles 'foo' and 'bar' are 
both active.

This patch adds support for bean profiles.  Bean profiles are 
activated by specifying the '-profiles' option to the 
UniversalSpringCell.  The argument is a comma-separated list of bean 
profiles to activate; there is no support for bean profiles containing 
a comma in the name.

The current method of allowing dCache configuration to alter which 
beans are instantiated is to split the Spring XML file into several 
files and to parametrise the XML filename with the corresponding 
dCache configured property.  Common definitions are moved to a 
separate file and included in each property-value-specific XML file.

With bean profiles, this is no longer needed.  Instead, the set of 
dCache configuration properties that affect which beans are loaded are 
mapped into corresponding set of bean profiles.  A single XML Spring 
file is used, with bean profiles controlling which beans to 
instantiate.

To demonstrate this, the gPlazma cell is modified to use bean 
profiles.  Two bean profiles are introduced: 'gplazma-1' and 
'gplazma-2', depending on which gPlazma version is requested.  Each 
profile defines a bean with the same id (the 'login-strategy') but if 
'gplazma-1' is active then GplazmaLoginStrategy is instantiated while 
Gplazma2LoginStrategy is instantiated if 'gplazma-2' bean profile is 
active.

Note that, under this scheme specifying both 'gplazma-1' and 
'gplazma-2' bean profiles at the same time or not specifying either 
bean profile will result in an invalid spring file.

Subsequent patches will update other spring cells to use bean 
profiles.

Target: trunk
Patch: http://rb.dcache.org/r/4223/
Acked-by: Tigran Mkrtchyan