webdav: work-around Java bug
Java JRE (both Sun and OpenJDK) contain a bug where, if the 
four-argument constructor is used and the path both contains a colon 
and the part before the colon is not a valid scheme part then the 
URI's construction will fail with URISyntaxException.

An example is a filename is "foo :bar".  This triggers the bug since 
scheme names may not contain spaces and Java misinterprets "foo " as 
a scheme name.

This bug has been reported:

    https://bugs.openjdk.java.net/show_bug.cgi?id=100223

The effect within dCache of this bug is that, if a directory contains 
a file or directory that triggers the bug (e.g. "foo :bar") then any 
attempt to view that directory will fail with a 500 Internal Error 
page being shown.

The work-around is to specify the "file" scheme.  This, in turn, 
requires that the path be absolute.  The patch prepends "/" to the 
path to ensure it is absolute.  In effect, this prepends "file:/" to 
the path.  The patch strips off "file:/" from the resulting string.

Target: trunk
Request: 2.1
Request: 2.0
Request: 1.9.12
Patch: http://rb.dcache.org/r/4236
Acked-by: Tigran Mkrtchyan