login service: catch bugs rather than letting them propagate to caller
The 'login service', typically run as the 'gplazma' service, provides 
a cell that other cells (typically doors) can query via the 
RemoteLoginStrategy.  This service wraps some LoginStrategy and passes 
the queries it receives onto this LoginStrategy.

The LoginStrategy (that the login service wraps) may have a bug that 
results in it throwing a RuntimeException.  These will be sent back to 
the calling cell (e.g., a door) and exposed as a CacheException.  This 
is undesirable as it introduces a disconnect between where the bug was 
and the effect of the bug.

This patch alters this behaviour: instead of propagating the 
exception, the bug is presented as a failed attempt at whichever 
operation was requested (login, map or reverseMap).  The bug is also 
logged inside the login service.

Target: trunk
Patch: http://rb.dcache.org/r/4230
Acked-by: Tigran Mkrtchyan
Acked-by: Gerd Behrmann