gplazma2: fix error behaviour
gPlazma2 uses a number of exceptions and errors to indicate problems.  
The exceptions are subclasses of RuntimeException (itself, a dubious 
choice) however, these exceptions are not caught appropriately, perhaps 
because of their unchecked nature.  There are three problems that this 
patch addresses:

First, gPlazma2 uses IllegalArgumentException to indicate that a plugin 
failed to instantiate.  gPlazma2 allows such exceptions to escape out 
of the GPlazma object.  The effect of an escaping 
IllegalArgumentException is that components (typically either doors or 
web-admin) that employ a RemoteLoginStrategy will receive an 
IllegalArgumentException for their call to the login method.  This 
exception is mostly unexpected and generally causes the component 
(door, webadmin, etc) to react as if there was a bug.

Second, gPlazma2 also uses a ParseException to indicate that there was 
a problem reading the gplazma.conf file.  This exception is a subclass 
of RuntimeException, which means it's unchecked.  The gPlazma code did 
not catch this exception, resulting in it propagating outside gPlazma.

Third, gPlazma2 uses FactoryConfigurationError to indicate it could not 
instantiate certain plugin classes.  This will kill the gPlazma cell; 
however, later code may choose to substitute a default behaviour or 
react differently.

This patch does the following to solve these problems.

It introduces PluginLoadingException, which is thrown by the loader 
infrastructure to signify a problem instantiating a plugin.

It modifies ParseException to be a checked exception.

It introduces FactoryConfigurationException: a checked exception that 
replaces FactoryConfigurationError.

It introduces GPlazmaInternalException as a base class for 
PluginLoadingException, ParseException and 
FactoryConfigurationException.

gPlazma2's behaviour is modified so that, if a GPlazmaInternalException 
is thrown during a login, map or reverseMap then all subsequent 
attempts to call login, map or reverseMap will fail with the same 
GPlazmaInternalException until a configuration update ('touch 
gplazma.conf') triggers re-initialising.

If a call to login, map or reverseMap result in 
GPlazmaInternalException then these methods will convert the exception 
to an appropriate form (login --> AuthenticationException, map & 
reverseMap -> NoSuchPrincipalException) with the error message 
describing the problem as an "internal gPlazma error" along with some 
details.

Target: trunk
Patch: http://rb.dcache.org/r/4229/
Acked-by: Tigran Mkrtchyan